# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZippiCRM API
  description: |
    Multi-tenant CRM system API for managing customers, products, orders, campaigns, and sales operations.

    ## Authentication
    This API uses session-based authentication. Users must first login via `/api/v1/auth/login` to obtain a session cookie.

    ## Multi-tenancy
    Most endpoints are tenant-scoped, requiring users to be authenticated and associated with a tenant.
    Super admins have access to all tenants.

    ## User Roles
    - **super_admin**: Full system access across all tenants
    - **manager**: Full access within their tenant(s)
    - **seller**: Limited access for sales operations within their tenant(s)
  version: 1.0.0
  contact:
    name: ZippiCRM Support
    email: admin@zippi.crm
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.zippi.crm
    description: Production server
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using cookies
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
      example:
        error: "Invalid credentials"

    # ==================== AUTH MODELS ====================
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Não autenticado"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Permissão insuficiente"
...
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Dados inválidos"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Erro interno do servidor"
paths:
...
  /api/v1/contact:
    post:
      summary: Register a public contact request
      description: |
        Public endpoint, rate limited to 5 requests per hour per IP. The payload
        is closed: unknown fields are rejected, so a caller cannot set the triage
        status. `consent` must be `true` and is recorded with the policy version.
        A non-empty `website` is a honeypot and the submission is discarded.
      operationId: createContactRequest
      tags:
        - Public
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/demo:
    post:
      summary: Register a public demo request
      description: |
        Public endpoint, rate limited to 5 requests per hour per IP. Same closed
        payload, mandatory recorded consent and honeypot as `/contact`.
      operationId: createDemoRequest
      tags:
        - Public
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...